Tests/LibWeb/Text/expected/Worker/Worker-crypto.txt

Type of self.crypto: [object Crypto]
